requirements.txt
aioredis==1.2.0
asgiref==3.2.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.2
Automat==0.7.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
dj-database-url==0.5.0
Django==2.2.5
django-bootstrap3==11.1.0
django-crispy-forms==1.7.2
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
iso8601==0.1.12
msgpack==0.6.1
ndg-httpsclient==0.5.1
paho-mqtt==1.4.0
pusher==2.1.4
pyasn1==0.4.7
pycparser==2.19
PyHamcrest==1.9.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.3
whitenoise==4.1.3
wiotp-sdk==0.9.5
zope.interface==4.6.0
django-background-tasks